optimize logs and event rendering with fragmenting
