+++ templates/index.html
@@ -208,7 +208,7 @@
     
     <script src="/static/js/fuse.js"></script>
     <script src="/static/js/navigation.js"></script>
-    <script src="/static/js/main.js"></script>
+    <script type="module" src="/static/js/main.js"></script>
     <script src="/static/js/metrics.js"></script>
 </body>
 </html>
